+++ lib/src/main/java/com/fabernovel/statefullayout/State.kt
@@ -5,7 +5,9 @@ import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.FrameLayout
+import androidx.annotation.AttrRes
 import androidx.annotation.LayoutRes
+import androidx.annotation.StyleRes
 import com.fabernovel.statefullayout.transitions.StateTransition
 import com.fabernovel.statefullayout.transitions.StateTransitionListener
 import com.fabernovel.statefullayout.transitions.StateTransitionProvider
@@ -39,24 +41,33 @@ class State : FrameLayout {
     constructor(
         context: Context,
         attrs: AttributeSet?
-    ) : this(context, attrs, R.style.Widget_Stateful_State)
+    ) : this(context, attrs, R.attr.stateStyle)
 
     constructor(
         context: Context,
         attrs: AttributeSet?,
-        defStyle: Int
-    ) : this(context, attrs, defStyle, R.attr.stateStyle)
+        @AttrRes defStyleAttr: Int
+    ) : this(context, attrs, defStyleAttr, R.style.Widget_Stateful_State)
 
     constructor(
         context: Context,
         attrs: AttributeSet?,
-        defStyle: Int,
-        defStyleRes: Int
-    ) : super(context, attrs, defStyle, defStyleRes) {
+        @AttrRes defStyleAttr: Int,
+        @StyleRes defStyleRes: Int
+    ) : super(context, attrs, defStyleAttr, defStyleRes) {
+        init(context, attrs, defStyleAttr, defStyleRes)
+    }
+
+    private fun init(
+        context: Context,
+        attrs: AttributeSet?,
+        @AttrRes defStyleAttr: Int,
+        @StyleRes defStyleRes: Int
+    ) {
         val array = context.obtainStyledAttributes(
             attrs,
             R.styleable.State,
-            defStyle,
+            defStyleAttr,
             defStyleRes
         )
         try {
@@ -112,10 +123,13 @@ class State : FrameLayout {
         }
     }
 
-    internal fun show(fallbackTransitionProvider: StateTransitionProvider? = null) {
+    internal fun show(
+        fallbackTransitionProvider: StateTransitionProvider? = null,
+        playTransition: Boolean = true
+    ) {
         currentTransition?.cancel()
         val transition = enterTransition ?: fallbackTransitionProvider?.get()
-        if (transition == null) {
+        if (transition == null || !playTransition) {
             visibility = View.VISIBLE
         } else {
             currentTransition = transition
@@ -127,10 +141,13 @@ class State : FrameLayout {
         }
     }
 
-    internal fun hide(fallbackTransitionProvider: StateTransitionProvider? = null) {
+    internal fun hide(
+        fallbackTransitionProvider: StateTransitionProvider? = null,
+        playTransition: Boolean = true
+    ) {
         currentTransition?.cancel()
         val transition = exitTransition ?: fallbackTransitionProvider?.get()
-        if (transition == null) {
+        if (transition == null || !playTransition) {
             visibility = View.GONE
         } else {
             currentTransition = transition